Split db.go, optimize profiles.go.
