+++ lxd/db.go
@@ -2,11 +2,7 @@ package main
 
 import (
 	"database/sql"
-	"encoding/hex"
 	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/mattn/go-sqlite3"
@@ -185,7 +181,7 @@ func createDb(db *sql.DB) (err error) {
 		}
 	}
 
-	err = createDefaultProfile(db)
+	err = dbProfileCreateDefault(db)
 
 	return err
 }
@@ -199,626 +195,6 @@ func dbGetSchema(db *sql.DB) (v int) {
 		return 0
 	}
 	return v
-}
-
-func dbUpdateFromV11(d *Daemon) error {
-	cNames, err := dbContainersList(d.db, cTypeSnapshot)
-	if err != nil {
-		return err
-	}
-
-	errors := 0
-
-	for _, cName := range cNames {
-		snappieces := strings.SplitN(cName, shared.SnapshotDelimiter, 2)
-		oldPath := shared.VarPath("containers", snappieces[0], "snapshots", snappieces[1])
-		newPath := shared.VarPath("snapshots", snappieces[0], snappieces[1])
-		if shared.PathExists(oldPath) && !shared.PathExists(newPath) {
-			shared.Log.Info(
-				"Moving snapshot",
-				log.Ctx{
-					"snapshot": cName,
-					"oldPath":  oldPath,
-					"newPath":  newPath})
-
-			// Rsync
-			// containers/<container>/snapshots/<snap0>
-			//   to
-			// snapshots/<container>/<snap0>
-			output, err := storageRsyncCopy(oldPath, newPath)
-			if err != nil {
-				shared.Log.Error(
-					"Failed rsync snapshot",
-					log.Ctx{
-						"snapshot": cName,
-						"output":   output,
-						"err":      err})
-				errors++
-				continue
-			}
-
-			// Remove containers/<container>/snapshots/<snap0>
-			if err := os.RemoveAll(oldPath); err != nil {
-				shared.Log.Error(
-					"Failed to remove the old snapshot path",
-					log.Ctx{
-						"snapshot": cName,
-						"oldPath":  oldPath,
-						"err":      err})
-
-				// Ignore this error.
-				// errors++
-				// continue
-			}
-
-			// Remove /var/lib/lxd/containers/<container>/snapshots
-			// if its empty.
-			cPathParent := filepath.Dir(oldPath)
-			if ok, _ := shared.PathIsEmpty(cPathParent); ok {
-				os.Remove(cPathParent)
-			}
-
-		} // if shared.PathExists(oldPath) && !shared.PathExists(newPath) {
-	} // for _, cName := range cNames {
-
-	// Refuse to start lxd if a rsync failed.
-	if errors > 0 {
-		return fmt.Errorf("Got errors while moving snapshots, see the log output.")
-	}
-
-	stmt := `
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err = d.db.Exec(stmt, 12)
-	return err
-
-}
-
-func dbUpdateFromV10(d *Daemon) error {
-	if shared.PathExists(shared.VarPath("lxc")) {
-		err := os.Rename(shared.VarPath("lxc"), shared.VarPath("containers"))
-		if err != nil {
-			return err
-		}
-
-		shared.Debugf("Restarting all the containers following directory rename.")
-		containersShutdown(d)
-		containersRestart(d)
-	}
-
-	stmt := `
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := d.db.Exec(stmt, 11)
-	return err
-}
-
-func dbUpdateFromV9(db *sql.DB) error {
-	stmt := `
-CREATE TABLE tmp (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type VARCHAR(255) NOT NULL default "none",
-    FOREIGN KEY (container_id) REFERENCES containers (id) ON DELETE CASCADE,
-    UNIQUE (container_id, name)
-);
-
-INSERT INTO tmp SELECT * FROM containers_devices;
-
-UPDATE containers_devices SET type=0 WHERE id IN (SELECT id FROM tmp WHERE type="none");
-UPDATE containers_devices SET type=1 WHERE id IN (SELECT id FROM tmp WHERE type="nic");
-UPDATE containers_devices SET type=2 WHERE id IN (SELECT id FROM tmp WHERE type="disk");
-UPDATE containers_devices SET type=3 WHERE id IN (SELECT id FROM tmp WHERE type="unix-char");
-UPDATE containers_devices SET type=4 WHERE id IN (SELECT id FROM tmp WHERE type="unix-block");
-
-DROP TABLE tmp;
-
-CREATE TABLE tmp (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    profile_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type VARCHAR(255) NOT NULL default "none",
-    FOREIGN KEY (profile_id) REFERENCES profiles (id) ON DELETE CASCADE,
-    UNIQUE (profile_id, name)
-);
-
-INSERT INTO tmp SELECT * FROM profiles_devices;
-
-UPDATE profiles_devices SET type=0 WHERE id IN (SELECT id FROM tmp WHERE type="none");
-UPDATE profiles_devices SET type=1 WHERE id IN (SELECT id FROM tmp WHERE type="nic");
-UPDATE profiles_devices SET type=2 WHERE id IN (SELECT id FROM tmp WHERE type="disk");
-UPDATE profiles_devices SET type=3 WHERE id IN (SELECT id FROM tmp WHERE type="unix-char");
-UPDATE profiles_devices SET type=4 WHERE id IN (SELECT id FROM tmp WHERE type="unix-block");
-
-DROP TABLE tmp;
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 10)
-	return err
-}
-
-func dbUpdateFromV8(db *sql.DB) error {
-	stmt := `
-UPDATE certificates SET fingerprint = replace(fingerprint, " ", "");
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 9)
-	return err
-}
-
-func dbUpdateFromV7(db *sql.DB) error {
-	stmt := `
-UPDATE config SET key='core.trust_password' WHERE key IN ('password', 'trust_password', 'trust-password', 'core.trust-password');
-DELETE FROM config WHERE key != 'core.trust_password';
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 8)
-	return err
-}
-
-func dbUpdateFromV6(db *sql.DB) error {
-	// This update recreates the schemas that need an ON DELETE CASCADE foreign
-	// key.
-	stmt := `
-PRAGMA foreign_keys=OFF; -- So that integrity doesn't get in the way for now
-
-CREATE TEMP TABLE tmp AS SELECT * FROM containers_config;
-DROP TABLE containers_config;
-CREATE TABLE IF NOT EXISTS containers_config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    FOREIGN KEY (container_id) REFERENCES containers (id) ON DELETE CASCADE,
-    UNIQUE (container_id, key)
-);
-INSERT INTO containers_config SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM containers_devices;
-DROP TABLE containers_devices;
-CREATE TABLE IF NOT EXISTS containers_devices (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type INTEGER NOT NULL default 0,
-    FOREIGN KEY (container_id) REFERENCES containers (id) ON DELETE CASCADE,
-    UNIQUE (container_id, name)
-);
-INSERT INTO containers_devices SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM containers_devices_config;
-DROP TABLE containers_devices_config;
-CREATE TABLE IF NOT EXISTS containers_devices_config (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_device_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    FOREIGN KEY (container_device_id) REFERENCES containers_devices (id) ON DELETE CASCADE,
-    UNIQUE (container_device_id, key)
-);
-INSERT INTO containers_devices_config SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM containers_profiles;
-DROP TABLE containers_profiles;
-CREATE TABLE IF NOT EXISTS containers_profiles (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    profile_id INTEGER NOT NULL,
-    apply_order INTEGER NOT NULL default 0,
-    UNIQUE (container_id, profile_id),
-    FOREIGN KEY (container_id) REFERENCES containers(id) ON DELETE CASCADE,
-    FOREIGN KEY (profile_id) REFERENCES profiles(id) ON DELETE CASCADE
-);
-INSERT INTO containers_profiles SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM images_aliases;
-DROP TABLE images_aliases;
-CREATE TABLE IF NOT EXISTS images_aliases (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    image_id INTEGER NOT NULL,
-    description VARCHAR(255),
-    FOREIGN KEY (image_id) REFERENCES images (id) ON DELETE CASCADE,
-    UNIQUE (name)
-);
-INSERT INTO images_aliases SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM images_properties;
-DROP TABLE images_properties;
-CREATE TABLE IF NOT EXISTS images_properties (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    image_id INTEGER NOT NULL,
-    type INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    FOREIGN KEY (image_id) REFERENCES images (id) ON DELETE CASCADE
-);
-INSERT INTO images_properties SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM profiles_config;
-DROP TABLE profiles_config;
-CREATE TABLE IF NOT EXISTS profiles_config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value VARCHAR(255),
-    UNIQUE (profile_id, key),
-    FOREIGN KEY (profile_id) REFERENCES profiles(id) ON DELETE CASCADE
-);
-INSERT INTO profiles_config SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM profiles_devices;
-DROP TABLE profiles_devices;
-CREATE TABLE IF NOT EXISTS profiles_devices (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type INTEGER NOT NULL default 0,
-    UNIQUE (profile_id, name),
-    FOREIGN KEY (profile_id) REFERENCES profiles (id) ON DELETE CASCADE
-);
-INSERT INTO profiles_devices SELECT * FROM tmp;
-DROP TABLE tmp;
-
-CREATE TEMP TABLE tmp AS SELECT * FROM profiles_devices_config;
-DROP TABLE profiles_devices_config;
-CREATE TABLE IF NOT EXISTS profiles_devices_config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_device_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    UNIQUE (profile_device_id, key),
-    FOREIGN KEY (profile_device_id) REFERENCES profiles_devices (id) ON DELETE CASCADE
-);
-INSERT INTO profiles_devices_config SELECT * FROM tmp;
-DROP TABLE tmp;
-
-PRAGMA foreign_keys=ON; -- Make sure we turn integrity checks back on.
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 7)
-	if err != nil {
-		return err
-	}
-
-	// Get the rows with broken foreign keys an nuke them
-	rows, err := db.Query("PRAGMA foreign_key_check;")
-	if err != nil {
-		return err
-	}
-
-	var tablestodelete []string
-	var rowidtodelete []int
-
-	defer rows.Close()
-	for rows.Next() {
-		var tablename string
-		var rowid int
-		var targetname string
-		var keynumber int
-
-		rows.Scan(&tablename, &rowid, &targetname, &keynumber)
-		tablestodelete = append(tablestodelete, tablename)
-		rowidtodelete = append(rowidtodelete, rowid)
-	}
-	rows.Close()
-
-	for i := range tablestodelete {
-		_, err = db.Exec(fmt.Sprintf("DELETE FROM %s WHERE rowid = %d;", tablestodelete[i], rowidtodelete[i]))
-		if err != nil {
-			return err
-		}
-	}
-
-	return err
-}
-
-func dbUpdateFromV5(db *sql.DB) error {
-	stmt := `
-ALTER TABLE containers ADD COLUMN power_state INTEGER NOT NULL DEFAULT 0;
-ALTER TABLE containers ADD COLUMN ephemeral INTEGER NOT NULL DEFAULT 0;
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 6)
-	return err
-}
-
-func dbUpdateFromV4(db *sql.DB) error {
-	stmt := `
-CREATE TABLE IF NOT EXISTS config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    UNIQUE (key)
-);
-INSERT INTO schema (version, updated_at) VALUES (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 5)
-	if err != nil {
-		return err
-	}
-
-	passfname := shared.VarPath("adminpwd")
-	passOut, err := os.Open(passfname)
-	oldPassword := ""
-	if err == nil {
-		defer passOut.Close()
-		buff := make([]byte, 96)
-		_, err = passOut.Read(buff)
-		if err != nil {
-			return err
-		}
-
-		oldPassword = hex.EncodeToString(buff)
-		stmt := `INSERT INTO config (key, value) VALUES ("core.trust_password", ?);`
-
-		_, err := db.Exec(stmt, oldPassword)
-		if err != nil {
-			return err
-		}
-
-		return os.Remove(passfname)
-	}
-
-	return nil
-}
-
-func dbUpdateFromV3(db *sql.DB) error {
-	err := createDefaultProfile(db)
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec(`INSERT INTO schema (version, updated_at) values (?, strftime("%s"));`, 4)
-	return err
-}
-
-func dbUpdateFromV2(db *sql.DB) error {
-	stmt := `
-CREATE TABLE IF NOT EXISTS containers_devices (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type INTEGER NOT NULL default 0,
-    FOREIGN KEY (container_id) REFERENCES containers (id) ON DELETE CASCADE,
-    UNIQUE (container_id, name)
-);
-CREATE TABLE IF NOT EXISTS containers_devices_config (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_device_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    FOREIGN KEY (container_device_id) REFERENCES containers_devices (id),
-    UNIQUE (container_device_id, key)
-);
-CREATE TABLE IF NOT EXISTS containers_profiles (
-    id INTEGER primary key AUTOINCREMENT NOT NULL,
-    container_id INTEGER NOT NULL,
-    profile_id INTEGER NOT NULL,
-    apply_order INTEGER NOT NULL default 0,
-    UNIQUE (container_id, profile_id),
-    FOREIGN KEY (container_id) REFERENCES containers(id) ON DELETE CASCADE,
-    FOREIGN KEY (profile_id) REFERENCES profiles(id) ON DELETE CASCADE
-);
-CREATE TABLE IF NOT EXISTS profiles (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    UNIQUE (name)
-);
-CREATE TABLE IF NOT EXISTS profiles_config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value VARCHAR(255),
-    UNIQUE (profile_id, key),
-    FOREIGN KEY (profile_id) REFERENCES profiles(id) ON DELETE CASCADE
-);
-CREATE TABLE IF NOT EXISTS profiles_devices (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_id INTEGER NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    type INTEGER NOT NULL default 0,
-    UNIQUE (profile_id, name),
-    FOREIGN KEY (profile_id) REFERENCES profiles (id) ON DELETE CASCADE
-);
-CREATE TABLE IF NOT EXISTS profiles_devices_config (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    profile_device_id INTEGER NOT NULL,
-    key VARCHAR(255) NOT NULL,
-    value TEXT,
-    UNIQUE (profile_device_id, key),
-    FOREIGN KEY (profile_device_id) REFERENCES profiles_devices (id)
-);
-INSERT INTO schema (version, updated_at) values (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 3)
-	return err
-}
-
-/* Yeah we can do htis in a more clever way */
-func dbUpdateFromV1(db *sql.DB) error {
-	// v1..v2 adds images aliases
-	stmt := `
-CREATE TABLE IF NOT EXISTS images_aliases (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    name VARCHAR(255) NOT NULL,
-    image_id INTEGER NOT NULL,
-    description VARCHAR(255),
-    FOREIGN KEY (image_id) REFERENCES images (id) ON DELETE CASCADE,
-    UNIQUE (name)
-);
-INSERT INTO schema (version, updated_at) values (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 2)
-	return err
-}
-
-func dbUpdateFromV0(db *sql.DB) error {
-	// v0..v1 adds schema table
-	stmt := `
-CREATE TABLE IF NOT EXISTS schema (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    version INTEGER NOT NULL,
-    updated_at DATETIME NOT NULL,
-    UNIQUE (version)
-);
-INSERT INTO schema (version, updated_at) values (?, strftime("%s"));`
-	_, err := db.Exec(stmt, 1)
-	return err
-}
-
-func dbUpdate(d *Daemon, prevVersion int) error {
-	db := d.db
-
-	if prevVersion < 0 || prevVersion > DB_CURRENT_VERSION {
-		return fmt.Errorf("Bad database version: %d\n", prevVersion)
-	}
-	if prevVersion == DB_CURRENT_VERSION {
-		return nil
-	}
-	var err error
-	if prevVersion < 1 {
-		err = dbUpdateFromV0(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 2 {
-		err = dbUpdateFromV1(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 3 {
-		err = dbUpdateFromV2(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 4 {
-		err = dbUpdateFromV3(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 5 {
-		err = dbUpdateFromV4(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 6 {
-		err = dbUpdateFromV5(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 7 {
-		err = dbUpdateFromV6(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 8 {
-		err = dbUpdateFromV7(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 9 {
-		err = dbUpdateFromV8(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 10 {
-		err = dbUpdateFromV9(db)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 11 {
-		err = dbUpdateFromV10(d)
-		if err != nil {
-			return err
-		}
-	}
-	if prevVersion < 12 {
-		err = dbUpdateFromV11(d)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func createDefaultProfile(db *sql.DB) error {
-	rows, err := dbQuery(db, "SELECT id FROM profiles WHERE name=?", "default")
-	if err != nil {
-		return err
-	}
-	defer rows.Close()
-	id := -1
-	for rows.Next() {
-		var xID int
-		rows.Scan(&xID)
-		id = xID
-	}
-	if id != -1 {
-		// default profile already exists
-		return nil
-	}
-
-	tx, err := dbBegin(db)
-	if err != nil {
-		return err
-	}
-	result, err := tx.Exec("INSERT INTO profiles (name) VALUES (?)", "default")
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	id64, err := result.LastInsertId()
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	id = int(id64)
-
-	result, err = tx.Exec(`INSERT INTO profiles_devices
-		(profile_id, name, type) VALUES (?, ?, ?)`,
-		id, "eth0", 1)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	id64, err = result.LastInsertId()
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	devID := int(id64)
-
-	_, err = tx.Exec(`INSERT INTO profiles_devices_config
-		(profile_device_id, key, value) VALUES (?, ?, ?)`,
-		devID, "nictype", "bridged")
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	/* TODO - analyze system to choose a bridge */
-	_, err = tx.Exec(`INSERT INTO profiles_devices_config
-		(profile_device_id, key, value) VALUES (?, ?, ?)`,
-		devID, "parent", "lxcbr0")
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 // Create a database connection object and return it.
@@ -863,377 +239,6 @@ func initDb(d *Daemon) (err error) {
 	path := shared.VarPath("lxd.db")
 	err = initializeDbObject(d, path)
 	return err
-}
-
-func dbDevicesAdd(tx *sql.Tx, w string, cID int, devices shared.Devices) error {
-	str1 := fmt.Sprintf("INSERT INTO %ss_devices (%s_id, name, type) VALUES (?, ?, ?)", w, w)
-	stmt1, err := tx.Prepare(str1)
-	if err != nil {
-		return err
-	}
-	defer stmt1.Close()
-	str2 := fmt.Sprintf("INSERT INTO %ss_devices_config (%s_device_id, key, value) VALUES (?, ?, ?)", w, w)
-	stmt2, err := tx.Prepare(str2)
-	if err != nil {
-		return err
-	}
-	defer stmt2.Close()
-	for k, v := range devices {
-		if !ValidDeviceType(v["type"]) {
-			return fmt.Errorf("Invalid device type %s\n", v["type"])
-		}
-		result, err := stmt1.Exec(cID, k, v["type"])
-		if err != nil {
-			return err
-		}
-		id64, err := result.LastInsertId()
-		if err != nil {
-			return fmt.Errorf("Error inserting device %s into database", k)
-		}
-		// TODO: is this really int64? we should fix it everywhere if so
-		id := int(id64)
-		for ck, cv := range v {
-			if ck == "type" {
-				continue
-			}
-			if !ValidDeviceConfig(v["type"], ck, cv) {
-				return fmt.Errorf("Invalid device config %s %s\n", ck, cv)
-			}
-			_, err = stmt2.Exec(id, ck, cv)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func dbDeviceConfigGet(db *sql.DB, id int, isprofile bool) (shared.Device, error) {
-	var query string
-	var key, value string
-	newdev := shared.Device{} // That's a map[string]string
-	inargs := []interface{}{id}
-	outfmt := []interface{}{key, value}
-
-	if isprofile {
-		query = `SELECT key, value FROM profiles_devices_config WHERE profile_device_id=?`
-	} else {
-		query = `SELECT key, value FROM containers_devices_config WHERE container_device_id=?`
-	}
-
-	results, err := dbQueryScan(db, query, inargs, outfmt)
-
-	if err != nil {
-		return newdev, err
-	}
-
-	for _, r := range results {
-		key = r[0].(string)
-		value = r[1].(string)
-		newdev[key] = value
-	}
-
-	return newdev, nil
-}
-
-func dbDevicesGet(db *sql.DB, qName string, isprofile bool) (shared.Devices, error) {
-	var q string
-	if isprofile {
-		q = `SELECT profiles_devices.id, profiles_devices.name, profiles_devices.type
-			FROM profiles_devices JOIN profiles
-			ON profiles_devices.profile_id = profiles.id
-   		WHERE profiles.name=?`
-	} else {
-		q = `SELECT containers_devices.id, containers_devices.name, containers_devices.type
-			FROM containers_devices JOIN containers
-			ON containers_devices.container_id = containers.id
-			WHERE containers.name=?`
-	}
-	var id, dtype int
-	var name, stype string
-	inargs := []interface{}{qName}
-	outfmt := []interface{}{id, name, dtype}
-	results, err := dbQueryScan(db, q, inargs, outfmt)
-	if err != nil {
-		return nil, err
-	}
-
-	devices := shared.Devices{}
-	for _, r := range results {
-		id = r[0].(int)
-		name = r[1].(string)
-		stype, err = dbDeviceTypeToString(r[2].(int))
-		if err != nil {
-			return nil, err
-		}
-		newdev, err := dbDeviceConfigGet(db, id, isprofile)
-		if err != nil {
-			return nil, err
-		}
-		newdev["type"] = stype
-		devices[name] = newdev
-	}
-
-	return devices, nil
-}
-
-// dbCertInfo is here to pass the certificates content
-// from the database around
-type dbCertInfo struct {
-	ID          int
-	Fingerprint string
-	Type        int
-	Name        string
-	Certificate string
-}
-
-// dbCertsGet returns all certificates from the DB as CertBaseInfo objects.
-func dbCertsGet(db *sql.DB) (certs []*dbCertInfo, err error) {
-	rows, err := dbQuery(
-		db,
-		"SELECT id, fingerprint, type, name, certificate FROM certificates",
-	)
-	if err != nil {
-		return certs, err
-	}
-
-	defer rows.Close()
-
-	for rows.Next() {
-		cert := new(dbCertInfo)
-		rows.Scan(
-			&cert.ID,
-			&cert.Fingerprint,
-			&cert.Type,
-			&cert.Name,
-			&cert.Certificate,
-		)
-		certs = append(certs, cert)
-	}
-
-	return certs, nil
-}
-
-// dbCertGet gets an CertBaseInfo object from the database.
-// The argument fingerprint will be queried with a LIKE query, means you can
-// pass a shortform and will get the full fingerprint.
-// There can never be more than one image with a given fingerprint, as it is
-// enforced by a UNIQUE constraint in the schema.
-func dbCertGet(db *sql.DB, fingerprint string) (cert *dbCertInfo, err error) {
-	cert = new(dbCertInfo)
-
-	inargs := []interface{}{fingerprint + "%"}
-	outfmt := []interface{}{
-		&cert.ID,
-		&cert.Fingerprint,
-		&cert.Type,
-		&cert.Name,
-		&cert.Certificate,
-	}
-
-	query := `
-		SELECT
-			id, fingerprint, type, name, certificate
-		FROM
-			certificates
-		WHERE fingerprint LIKE ?`
-
-	if err = dbQueryRowScan(db, query, inargs, outfmt); err != nil {
-		return nil, err
-	}
-
-	return cert, err
-}
-
-// dbCertSave stores a CertBaseInfo object in the db,
-// it will ignore the ID field from the dbCertInfo.
-func dbCertSave(db *sql.DB, cert *dbCertInfo) error {
-	tx, err := dbBegin(db)
-	if err != nil {
-		return err
-	}
-	stmt, err := tx.Prepare(`
-			INSERT INTO certificates (
-				fingerprint,
-				type,
-				name,
-				certificate
-			) VALUES (?, ?, ?, ?)`,
-	)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-	defer stmt.Close()
-	_, err = stmt.Exec(
-		cert.Fingerprint,
-		cert.Type,
-		cert.Name,
-		cert.Certificate,
-	)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	return txCommit(tx)
-}
-
-// dbCertDelete deletes a certificate from the db.
-func dbCertDelete(db *sql.DB, fingerprint string) error {
-	_, err := dbExec(
-		db,
-		"DELETE FROM certificates WHERE fingerprint=?",
-		fingerprint,
-	)
-
-	return err
-}
-
-// dbImageGet gets an ImageBaseInfo object from the database.
-// The argument fingerprint will be queried with a LIKE query, means you can
-// pass a shortform and will get the full fingerprint.
-// There can never be more than one image with a given fingerprint, as it is
-// enforced by a UNIQUE constraint in the schema.
-func dbImageGet(db *sql.DB, fingerprint string, public bool) (*shared.ImageBaseInfo, error) {
-	var err error
-	var create, expire, upload *time.Time // These hold the db-returned times
-
-	// The object we'll actually return
-	image := new(shared.ImageBaseInfo)
-
-	// These two humongous things will be filled by the call to DbQueryRowScan
-	inargs := []interface{}{fingerprint + "%"}
-	outfmt := []interface{}{&image.Id, &image.Fingerprint, &image.Filename,
-		&image.Size, &image.Public, &image.Architecture,
-		&create, &expire, &upload}
-
-	query := `
-        SELECT
-            id, fingerprint, filename, size, public, architecture,
-            creation_date, expiry_date, upload_date
-        FROM
-            images
-        WHERE fingerprint like ?`
-
-	if public {
-		query = query + " AND public=1"
-	}
-
-	err = dbQueryRowScan(db, query, inargs, outfmt)
-
-	if err != nil {
-		return nil, err // Likely: there are no rows for this fingerprint
-	}
-
-	// Some of the dates can be nil in the DB, let's process them.
-	if create != nil {
-		image.CreationDate = create.Unix()
-	} else {
-		image.CreationDate = 0
-	}
-	if expire != nil {
-		image.ExpiryDate = expire.Unix()
-	} else {
-		image.ExpiryDate = 0
-	}
-	// The upload date is enforced by NOT NULL in the schema, so it can never be nil.
-	image.UploadDate = upload.Unix()
-
-	return image, nil
-}
-
-func dbImageDelete(db *sql.DB, id int) error {
-	tx, err := dbBegin(db)
-	if err != nil {
-		return err
-	}
-
-	_, _ = tx.Exec("DELETE FROM images_aliases WHERE image_id=?", id)
-	_, _ = tx.Exec("DELETE FROM images_properties WHERE image_id?", id)
-	_, _ = tx.Exec("DELETE FROM images WHERE id=?", id)
-
-	if err := txCommit(tx); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Get an image's fingerprint for a given alias name.
-func dbImageAliasGet(db *sql.DB, name string) (fingerprint string, err error) {
-	q := `
-        SELECT
-            fingerprint
-        FROM images AS i JOIN images_aliases AS a
-        ON a.image_id == i.id
-        WHERE name=?`
-
-	inargs := []interface{}{name}
-	outfmt := []interface{}{&fingerprint}
-
-	err = dbQueryRowScan(db, q, inargs, outfmt)
-
-	if err == sql.ErrNoRows {
-		return "", NoSuchObjectError
-	}
-	if err != nil {
-		return "", err
-	}
-	return fingerprint, nil
-}
-
-// Insert an alias into the database.
-func dbImageAliasAdd(db *sql.DB, name string, imageID int, desc string) error {
-	stmt := `INSERT into images_aliases (name, image_id, description) values (?, ?, ?)`
-	_, err := dbExec(db, stmt, name, imageID, desc)
-	return err
-}
-
-// Get the profile configuration map from the DB
-func dbProfileConfigGet(db *sql.DB, name string) (map[string]string, error) {
-	var key, value string
-	query := `
-        SELECT
-            key, value
-        FROM profiles_config
-        JOIN profiles ON profiles_config.profile_id=profiles.id
-		WHERE name=?`
-	inargs := []interface{}{name}
-	outfmt := []interface{}{key, value}
-	results, err := dbQueryScan(db, query, inargs, outfmt)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to get profile '%s'", name)
-	}
-
-	if len(results) == 0 {
-		/*
-		 * If we didn't get any rows here, let's check to make sure the
-		 * profile really exists; if it doesn't, let's send back a 404.
-		 */
-		query := "SELECT id FROM profiles WHERE name=?"
-		var id int
-		results, err := dbQueryScan(db, query, []interface{}{name}, []interface{}{id})
-		if err != nil {
-			return nil, err
-		}
-
-		if len(results) == 0 {
-			return nil, NoSuchObjectError
-		}
-	}
-
-	config := map[string]string{}
-
-	for _, r := range results {
-		key = r[0].(string)
-		value = r[1].(string)
-
-		config[key] = value
-	}
-
-	return config, nil
 }
 
 func isDbLockedError(err error) bool {
